fix: add report path prefix to datacatalog
AWS recommends to have one CUR in one bucket. If it is needed to have multiple CUR exported to the
same bucket,  different report path prefixes should be used. In order to not crawl to much data, the
report path prefix  is added to the datacatalog to keep the effort of crawling to a minimum.

fix #7